requirements.txt
asgiref==3.4.1
Django==3.2.9
django-allauth==0.41.0
django-embed-video==1.4.0
oauthlib==3.1.1
Pillow==8.4.0
pylint-django==2.4.4
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2021.3
requests-oauthlib==1.3.0
sqlparse==0.4.2
